Move the existing webhooks from v1alpha1 to v1alpha2
