+++ api/v1alpha2/workflow_webhook.go
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package v1alpha1
+package v1alpha2
 
 import (
 	"context"
@@ -51,7 +51,7 @@ func (w *Workflow) SetupWebhookWithManager(mgr ctrl.Manager) error {
 		Complete()
 }
 
-//+kubebuilder:webhook:path=/mutate-dws-cray-hpe-com-v1alpha1-workflow,mutating=true,failurePolicy=fail,sideEffects=None,groups=dws.cray.hpe.com,resources=workflows,verbs=create,versions=v1alpha1,name=mworkflow.kb.io,admissionReviewVersions={v1,v1beta1}
+//+kubebuilder:webhook:path=/mutate-dws-cray-hpe-com-v1alpha2-workflow,mutating=true,failurePolicy=fail,sideEffects=None,groups=dws.cray.hpe.com,resources=workflows,verbs=create,versions=v1alpha2,name=mworkflow.kb.io,admissionReviewVersions={v1,v1beta1}
 
 var _ webhook.Defaulter = &Workflow{}
 
@@ -70,7 +70,7 @@ func (w *Workflow) Default() {
 }
 
 // TODO(user): change verbs to "verbs=create;update;delete" if you want to enable deletion validation.
-//+kubebuilder:webhook:path=/validate-dws-cray-hpe-com-v1alpha1-workflow,mutating=false,failurePolicy=fail,sideEffects=None,groups=dws.cray.hpe.com,resources=workflows,verbs=create;update,versions=v1alpha1,name=vworkflow.kb.io,admissionReviewVersions={v1,v1beta1}
+//+kubebuilder:webhook:path=/validate-dws-cray-hpe-com-v1alpha2-workflow,mutating=false,failurePolicy=fail,sideEffects=None,groups=dws.cray.hpe.com,resources=workflows,verbs=create;update,versions=v1alpha2,name=vworkflow.kb.io,admissionReviewVersions={v1,v1beta1}
 
 var _ webhook.Validator = &Workflow{}
