+++ api/v1alpha2/workflow_webhook_test.go
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package v1alpha1
+package v1alpha2
 
 import (
 	"context"
